+++ src/controllers/CartController.php
@@ -46,7 +46,7 @@ class CartController extends BaseFrontEndController
     /**
      * @var User|null
      */
-    protected ?User $_currentUser;
+    protected ?User $_currentUser = null;
 
     /**
      * @throws InvalidConfigException
@@ -94,6 +94,11 @@ class CartController extends BaseFrontEndController
         // When we are about to update the cart, we consider it a real cart at this point, and want to actually create it in the DB.
         $this->_cart = $this->_getCart(true);
 
+        // Can clear line items when updating the cart
+        if (($clearCart = $this->request->getParam('clearLineItems')) !== null) {
+            $this->_cart->setLineItems([]);
+        }
+
         // Can clear notices when updating the cart
         if ($this->request->getParam('clearNotices') !== null) {
             $this->_cart->clearNotices();
@@ -109,7 +114,7 @@ class CartController extends BaseFrontEndController
             $qty = (int)$this->request->getParam('qty', 1);
 
             if ($qty > 0) {
-                $lineItem = Plugin::getInstance()->getLineItems()->resolveLineItem($this->_cart->id, $purchasableId, $options);
+                $lineItem = Plugin::getInstance()->getLineItems()->resolveLineItem($this->_cart, $purchasableId, $options);
 
                 // New line items already have a qty of one.
                 if ($lineItem->id) {
@@ -155,7 +160,7 @@ class CartController extends BaseFrontEndController
 
                 // Ignore zero value qty for multi-add forms https://github.com/craftcms/commerce/issues/330#issuecomment-384533139
                 if ($purchasable['qty'] > 0) {
-                    $lineItem = Plugin::getInstance()->getLineItems()->resolveLineItem($this->_cart->id, $purchasable['id'], $purchasable['options']);
+                    $lineItem = Plugin::getInstance()->getLineItems()->resolveLineItem($this->_cart, $purchasable['id'], $purchasable['options']);
 
                     // New line items already have a qty of one.
                     if ($lineItem->id) {