+++ src/elements/Order.php
@@ -22,7 +22,6 @@ use craft\commerce\errors\CurrencyException;
 use craft\commerce\errors\OrderStatusException;
 use craft\commerce\events\AddLineItemEvent;
 use craft\commerce\events\LineItemEvent;
-use craft\commerce\exports\Raw;
 use craft\commerce\helpers\Currency;
 use craft\commerce\helpers\Order as OrderHelper;
 use craft\commerce\models\Address;
@@ -760,9 +759,9 @@ class Order extends Element
     public bool $estimatedBillingSameAsShipping = false;
 
     /**
-     * @var string Shipping Method Handle
+     * @var string|null Shipping Method Handle
      */
-    public string $shippingMethodHandle = '';
+    public ?string $shippingMethodHandle = '';
 
     /**
      * @var string|null Shipping Method Name
@@ -1077,15 +1076,19 @@ class Order extends Element
      */
     public function init(): void
     {
-        // Set default addresses on the order
+        // Set default addresseses
         if (!$this->isCompleted && Plugin::getInstance()->getSettings()->autoSetNewCartAddresses) {
-            $hasPrimaryShippingAddress = !$this->shippingAddressId && $this->getCustomer() && $this->getCustomer()->primaryShippingAddressId;
-            if ($hasPrimaryShippingAddress && ($shippingAddress = Plugin::getInstance()->getAddresses()->getAddressByIdAndCustomerId($this->getCustomer()->primaryShippingAddressId, $this->customerId))) {
-                $this->setShippingAddress($shippingAddress);
+            if (!$this->shippingAddressId) {
+                $hasPrimaryShippingAddress = $this->getCustomer() && $this->getCustomer()->primaryShippingAddressId;
+                if ($hasPrimaryShippingAddress && ($shippingAddress = Plugin::getInstance()->getAddresses()->getAddressByIdAndCustomerId($this->getCustomer()->primaryShippingAddressId, $this->customerId))) {
+                    $this->setShippingAddress($shippingAddress);
+                }
             }
-            $hasPrimaryBillingAddress = !$this->billingAddressId && $this->getCustomer() && $this->getCustomer()->primaryBillingAddressId;
-            if ($hasPrimaryBillingAddress && ($billingAddress = Plugin::getInstance()->getAddresses()->getAddressByIdAndCustomerId($this->getCustomer()->primaryBillingAddressId, $this->customerId))) {
-                $this->setBillingAddress($billingAddress);
+            if (!$this->billingAddressId) {
+                $hasPrimaryBillingAddress = $this->getCustomer() && $this->getCustomer()->primaryBillingAddressId;
+                if ($hasPrimaryBillingAddress && ($billingAddress = Plugin::getInstance()->getAddresses()->getAddressByIdAndCustomerId($this->getCustomer()->primaryBillingAddressId, $this->customerId))) {
+                    $this->setBillingAddress($billingAddress);
+                }
             }
         }
 
@@ -1116,6 +1119,13 @@ class Order extends Element
             } else {
                 $this->setRecalculationMode(self::RECALCULATION_MODE_ALL);
             }
+        }
+
+        // Sets a default shipping method
+        // Leave this as the last one inside init(), as shipping rules will need access the above default that are set (like currency).
+        if (!$this->shippingMethodHandle && !$this->isCompleted && Plugin::getInstance()->getSettings()->autoSetCartShippingMethodOption) {
+            $availableMethodOptions = $this->getAvailableShippingMethodOptions();
+            $this->shippingMethodHandle = ArrayHelper::firstKey($availableMethodOptions);
         }
 
         parent::init();
@@ -1364,7 +1374,6 @@ class Order extends Element
     public function extraFields(): array
     {
         $names = parent::extraFields();
-        $names[] = 'matchingShippingMethods';
         $names[] = 'availableShippingMethodOptions';
         $names[] = 'adjustments';
         $names[] = 'billingAddress';
@@ -1707,6 +1716,21 @@ class Order extends Element
         }
 
         if ($this->getRecalculationMode() == self::RECALCULATION_MODE_ALL) {
+
+            // Make sure we set a default shipping method option
+            if (!$this->isCompleted && Plugin::getInstance()->getSettings()->autoSetCartShippingMethodOption) {
+                $availableMethodOptions = $this->getAvailableShippingMethodOptions();
+                if (!$this->shippingMethodHandle || !isset($availableMethodOptions[$this->shippingMethodHandle])) {
+                    $this->shippingMethodHandle = ArrayHelper::firstKey($availableMethodOptions);
+                }
+            }
+
+            if (!$this->shippingMethodHandle) {
+                $this->shippingMethodName = null;
+            } else if ($shippingMethod = $this->getShippingMethod()) {
+                $this->shippingMethodName = $shippingMethod->getName();
+            }
+
             $lineItemRemoved = false;
             foreach ($this->getLineItems() as $item) {
                 $originalSalePrice = $item->getSalePrice();
@@ -1835,22 +1859,6 @@ class Order extends Element
         }
 
         return $options;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function beforeSave(bool $isNew): bool
-    {
-        if (null === $this->shippingMethodHandle) {
-            // Reset shipping method name if there is no handle
-            $this->shippingMethodName = null;
-        } else if ($this->shippingMethodHandle && $shippingMethod = $this->getShippingMethod()) {
-            // Update shipping method name if there is a handle and we can retrieve the method
-            $this->shippingMethodName = $shippingMethod->name;
-        }
-
-        return parent::beforeSave($isNew);
     }
 
     /**
@@ -2695,7 +2703,7 @@ class Order extends Element
      */
     public function getAdjustments(): ?array
     {
-        if (null !== $this->_orderAdjustments) {
+        if (isset($this->_orderAdjustments)) {
             return $this->_orderAdjustments;
         }
 
@@ -2772,7 +2780,7 @@ class Order extends Element
      */
     public function getShippingAddress(): ?Address
     {
-        if (null === $this->_shippingAddress && $this->shippingAddressId) {
+        if (!isset($this->_shippingAddress) && $this->shippingAddressId) {
             $this->_shippingAddress = Plugin::getInstance()->getAddresses()->getAddressById($this->shippingAddressId);
         }
 
@@ -2806,7 +2814,7 @@ class Order extends Element
         $this->_shippingAddress = $address;
 
         // When we are setting an address we need to keep them in sync if they have the same ID
-        if (null !== $this->shippingAddressId && null !== $this->billingAddressId && $this->billingAddressId === $this->shippingAddressId) {
+        if (isset($this->shippingAddressId) && isset($this->billingAddressId) && $this->billingAddressId === $this->shippingAddressId) {
             $this->_billingAddress = $this->_shippingAddress;
         }
     }
@@ -2826,7 +2834,7 @@ class Order extends Element
      */
     public function getEstimatedShippingAddress(): ?Address
     {
-        if (null === $this->_estimatedShippingAddress && $this->estimatedShippingAddressId) {
+        if (!isset($this->_estimatedShippingAddress) && $this->estimatedShippingAddressId) {
             $this->_estimatedShippingAddress = Plugin::getInstance()->getAddresses()->getAddressById($this->estimatedShippingAddressId);
         }
 
@@ -2864,7 +2872,7 @@ class Order extends Element
      */
     public function getBillingAddress(): ?Address
     {
-        if (null === $this->_billingAddress && $this->billingAddressId) {
+        if (!isset($this->_billingAddress) && $this->billingAddressId) {
             $this->_billingAddress = Plugin::getInstance()->getAddresses()->getAddressById($this->billingAddressId);
         }
 
@@ -2898,7 +2906,7 @@ class Order extends Element
         $this->_billingAddress = $address;
 
         // When we are setting an address we need to keep them in sync if they have the same ID
-        if (null !== $this->shippingAddressId && null !== $this->billingAddressId && $this->shippingAddressId === $this->billingAddressId) {
+        if (isset($this->shippingAddressId) && isset($this->billingAddressId) && $this->shippingAddressId === $this->billingAddressId) {
             $this->_shippingAddress = $this->_billingAddress;
         }
     }
@@ -2918,7 +2926,7 @@ class Order extends Element
      */
     public function getEstimatedBillingAddress(): ?Address
     {
-        if (null === $this->_estimatedBillingAddress && $this->estimatedBillingAddressId) {
+        if (!isset($this->_estimatedBillingAddress) && $this->estimatedBillingAddressId) {
             $this->_estimatedBillingAddress = Plugin::getInstance()->getAddresses()->getAddressById($this->estimatedBillingAddressId);
         }