+++ src/models/PaymentSource.php
@@ -59,12 +59,12 @@ class PaymentSource extends Model
     /**
      * @var User|null $_user
      */
-    private ?User $_user;
+    private ?User $_user = null;
 
     /**
      * @var GatewayInterface|null $_gateway
      */
-    private ?GatewayInterface $_gateway;
+    private ?GatewayInterface $_gateway = null;
 
 
     /**