+++ src/services/Sales.php
@@ -186,7 +186,7 @@ class Sales extends Component
      */
     public function getAllSales(): array
     {
-        if (null === $this->_allSales) {
+        if (!isset($this->_allSales)) {
             $sales = (new Query())->select([
                 'sales.id',
                 'sales.name',
@@ -252,49 +252,6 @@ class Sales extends Component
         }
 
         return $this->_allSales;
-    }
-
-    /**
-     * Populates a sale's relations.
-     *
-     * @param Sale $sale
-     * @deprecated in 3.2.0. No longer required as IDs are populated when retrieving the sale using the service.
-     * // TODO Removed when completing #COM-58
-     */
-    public function populateSaleRelations(Sale $sale): void
-    {
-        $rows = (new Query())->select(
-            'sp.purchasableId,
-            spt.categoryId,
-            sug.userGroupId')
-            ->from(Table::SALES . ' sales')
-            ->leftJoin(Table::SALE_PURCHASABLES . ' sp', '[[sp.saleId]]=[[sales.id]]')
-            ->leftJoin(Table::SALE_CATEGORIES . ' spt', '[[spt.saleId]]=[[sales.id]]')
-            ->leftJoin(Table::SALE_USERGROUPS . ' sug', '[[sug.saleId]]=[[sales.id]]')
-            ->where(['sales.id' => $sale->id])
-            ->all();
-
-        $purchasableIds = [];
-        $categoryIds = [];
-        $userGroupIds = [];
-
-        foreach ($rows as $row) {
-            if ($row['purchasableId']) {
-                $purchasableIds[] = $row['purchasableId'];
-            }
-
-            if ($row['categoryId']) {
-                $categoryIds[] = $row['categoryId'];
-            }
-
-            if ($row['userGroupId']) {
-                $userGroupIds[] = $row['userGroupId'];
-            }
-        }
-
-        $sale->setPurchasableIds($purchasableIds);
-        $sale->setCategoryIds($categoryIds);
-        $sale->setUserGroupIds($userGroupIds);
     }
 
     /**
@@ -404,7 +361,7 @@ class Sales extends Component
         $salePrice = ($originalPrice + $takeOffAmount);
 
         // A newPrice has been set so use it.
-        if (null !== $newPrice) {
+        if ($newPrice !== null) {
             $salePrice = $newPrice;
         }
 
@@ -477,7 +434,7 @@ class Sales extends Component
 
             if (!$sale->allGroups) {
                 // We must pass a real user to getCurrentUserGroupIds, otherwise the current user is used.
-                if (null === $user) {
+                if ($user === null) {
                     return false;
                 }
                 // User groups of the order's user
@@ -700,7 +657,7 @@ class Sales extends Component
      */
     private function _getAllEnabledSales(): array
     {
-        if (null === $this->_allActiveSales) {
+        if (!isset($this->_allActiveSales)) {
             $sales = $this->getAllSales();
             $activeSales = [];
             foreach ($sales as $sale) {