+++ tests/unit/services/LineItemsTest.php
@@ -8,6 +8,7 @@
 namespace craftcommercetests\unit\services;
 
 use Codeception\Test\Unit;
+use craft\commerce\elements\Order;
 use craft\commerce\elements\Variant;
 use craft\commerce\models\LineItem;
 use craft\commerce\Plugin;
@@ -73,11 +74,11 @@ class LineItemsTest extends Unit
 
     public function testResolveLineItemExisting(): void
     {
+        /** @var Order $order */
         $order = $this->fixtureData->getElement('completed-new');
-        /** @var LineItem $orderLineItem */
         $orderLineItem = $order->getLineItems()[0];
 
-        $resolvedLineItem = $this->service->resolveLineItem($order->id, $orderLineItem->purchasableId, $orderLineItem->getOptions());
+        $resolvedLineItem = $this->service->resolveLineItem($order, $orderLineItem->purchasableId, $orderLineItem->getOptions());
 
         self::assertInstanceOf(LineItem::class, $resolvedLineItem);
         // Test that resolving line items without saving is consistent
@@ -90,10 +91,23 @@ class LineItemsTest extends Unit
 
     public function testResolveLineItemNew(): void
     {
-        $lineItem = $this->fixtureData->getElement('completed-new')->getLineItems()[1];
+        /** @var Order $order */
+        $order = $this->fixtureData->getElement('completed-new');
+        $lineItem = $order->getLineItems()[1];
         $variant = Variant::find()->id($lineItem->purchasableId)->one();
 
-        $resolvedLineItem = $this->service->resolveLineItem($this->fixtureData->getElement('completed-shipped')->id, $lineItem->purchasableId, $lineItem->getOptions());
+        $resolvedLineItem = $this->service->resolveLineItem($order, $lineItem->purchasableId, $lineItem->getOptions());
+
+        self::assertInstanceOf(LineItem::class, $resolvedLineItem);
+        self::assertEquals($variant->getPrice(), $resolvedLineItem->getPrice());
+    }
+
+    public function testResolveLineItemUnsavedOrder(): void
+    {
+        $order = new Order();
+        $variant = Variant::find()->sku('hct-blue')->one();
+
+        $resolvedLineItem = $this->service->resolveLineItem($order, $variant->id, ['giftWrapped' => 'no']);
 
         self::assertInstanceOf(LineItem::class, $resolvedLineItem);
         self::assertEquals($variant->getPrice(), $resolvedLineItem->getPrice());
@@ -110,10 +124,12 @@ class LineItemsTest extends Unit
 
     public function testCreateLineItem(): void
     {
-        $lineItem = $this->fixtureData->getElement('completed-new')->getLineItems()[0];
+        /** @var Order $order */
+        $order = $this->fixtureData->getElement('completed-new');
+        $lineItem = $order->getLineItems()[0];
         $qty = 4;
         $note = 'My note';
-        $lineItem = $this->service->createLineItem($this->fixtureData->getElement('completed-new')->id, $lineItem->purchasableId, $lineItem->options, $qty, $note);
+        $lineItem = $this->service->createLineItem($order, $lineItem->purchasableId, $lineItem->options, $qty, $note);
 
         self::assertInstanceOf(LineItem::class, $lineItem);
         self::assertEquals($this->fixtureData->getElement('completed-new')->id, $lineItem->orderId);